Add GetModelContext command and remove old project info
