+++ Commands/GetModelContextCommand.cs
@@ -1,0 +1,93 @@
+// GetModelContextCommand.cs - retrieves model metadata and project parameters
+using Autodesk.Revit.DB;
+using Autodesk.Revit.UI;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+public class GetModelContextCommand : ICommand
+{
+    public Dictionary<string, object> Execute(UIApplication app, Dictionary<string, string> input)
+    {
+        var response = new Dictionary<string, object>();
+        var doc = app?.ActiveUIDocument?.Document;
+
+        if (doc == null)
+        {
+            response["status"] = "error";
+            response["message"] = "No active Revit document.";
+            return response;
+        }
+
+        try
+        {
+            // project information parameters
+            var projectInfo = doc.ProjectInformation;
+            var info = new Dictionary<string, string>();
+            foreach (Parameter p in projectInfo.Parameters)
+            {
+                if (p == null || string.IsNullOrEmpty(p.Definition?.Name)) continue;
+                string val = string.Empty;
+                switch (p.StorageType)
+                {
+                    case StorageType.String:
+                        val = p.AsString();
+                        break;
+                    case StorageType.Integer:
+                        val = p.AsInteger().ToString();
+                        break;
+                    case StorageType.Double:
+                        val = p.AsDouble().ToString();
+                        break;
+                    case StorageType.ElementId:
+                        val = p.AsElementId().IntegerValue.ToString();
+                        break;
+                }
+                info[p.Definition.Name] = val;
+            }
+
+            // project parameters
+            var bindingMap = doc.ParameterBindings;
+            var iterator = bindingMap.ForwardIterator();
+            var parameters = new List<Dictionary<string, object>>();
+
+            iterator.Reset();
+            while (iterator.MoveNext())
+            {
+                Definition definition = iterator.Key;
+                ElementBinding binding = iterator.Current as ElementBinding;
+
+                if (definition == null || binding == null) continue;
+
+                var paramData = new Dictionary<string, object>();
+                paramData["name"] = definition.Name;
+                paramData["parameter_type"] = definition.ParameterGroup.ToString();
+                paramData["unit_type_id"] = definition.GetDataType()?.TypeId?.ToString() ?? string.Empty;
+                paramData["binding_type"] = binding is InstanceBinding ? "Instance" : "Type";
+
+                var categories = new List<string>();
+                foreach (Category cat in binding.Categories)
+                {
+                    if (cat != null)
+                        categories.Add(cat.Name);
+                }
+                paramData["categories"] = categories;
+
+                parameters.Add(paramData);
+            }
+
+            response["status"] = "success";
+            response["model_name"] = doc.Title;
+            response["last_saved"] = System.IO.File.GetLastWriteTime(doc.PathName).ToString("yyyy-MM-ddTHH:mm:ss");
+            response["project_info"] = info;
+            response["project_parameters"] = parameters;
+        }
+        catch (Exception ex)
+        {
+            response["status"] = "error";
+            response["message"] = ex.Message;
+        }
+
+        return response;
+    }
+}
